final_project/requirements.txt
torch>=1.7.0
torchvision>=0.8.0
transformers>=4.30.0
numpy>=1.19.0
Pillow>=8.0.0
tqdm>=4.50.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
pandas>=1.1.0
nltk>=3.6.0
rouge-score>=0.0.4
pycocoevalcap>=1.2
tensorboard>=2.4.0
wandb>=0.12.0
opencv-python>=4.5.0 